Refactor plugin into PSR-4 structure with autoloading
